refactor(buttongroup): refactored buttongroup to namespace pattern
