+++ lib/src/Accordion/index.ts
@@ -1,1 +1,2 @@
 export * from "./Accordion";
+export { AccordionItemProps } from "./AccordionItem";
