+++ lib/src/Breadcrumb/index.ts
@@ -1,1 +1,2 @@
 export * from "./Breadcrumb";
+export { BreadcrumbItemProps } from "./BreadcrumbItem";
