+++ lib/src/Button/index.ts
@@ -1,1 +1,2 @@
 export * from "./Button";
+export { ButtonGroupProps, ButtonGroupSizes } from "./ButtonGroup";
