--- lib/src/ButtonGroup/index.ts
@@ -1,1 +1,0 @@
-export * from "./ButtonGroup";
